Fixed a couple of bugs in mt.iCIR and irserver.js
IR server now returns the correct time. The wrong time was returned at
certain points during the year because new Date() parses the date as a
US format, whereas the AutoExport function seems to use local formatting
(behaviour is not documented). So now the datetime is extracted from the
filename which is in more of an ISO format.